chore: update storybook coverage scripts and documentation
