--- scripts/storybook/index.ts
@@ -1,21 +1,0 @@
-/**
- * Storybook Component Inventory System
- *
- * This module provides tools for scanning, analyzing, and reporting
- * on Storybook story coverage across the codebase.
- *
- * Entry points:
- * - inventory.ts: Main scanner that generates component inventory
- * - coverage-report.ts: Generates detailed coverage reports
- *
- * Usage:
- *   yarn inventory         # Run component inventory
- *   yarn coverage-report   # Generate coverage report
- */
-
-export * from './types'
-export * from './scanner'
-export * from './coverage'
-export * from './dependencies'
-export * from './priority'
-export * from './family'
